Adjust some methods
The class GameObject was marked as abstract,  however, there was the need to instantiate it later.

The animationCallback attribute was set to be null in the Enemy class for there is no need for it.

Since the GameObject class is not abstract anymore, it does not need to be overridden in the child classes.