Implement the reset and game loop logics
The reset logic was implemented inside the function resetGame(), and it is responsible for reseting the positions of all the sprites on the screen.

The game loop logic is where the main logic of the game is built. Inside of it, all the sprites will have their states updated, and possible collisions will be detected. Also from this piece of code, the command to rebuild the game screen will be deployed. It is inside the function gameLoop().